chore: auto close menu on click
