+++ src/components/Channel/NewVideoTrigger.tsx
@@ -1,9 +1,9 @@
 import { Button } from '@components/UIElements/Button'
-import Popover from '@components/UIElements/Popover'
+import DropMenu, { NextLink } from '@components/UIElements/DropMenu'
 import Tooltip from '@components/UIElements/Tooltip'
+import { Menu } from '@headlessui/react'
 import usePersistStore from '@lib/store/persist'
 import { UPLOAD } from '@utils/url-path'
-import Link from 'next/link'
 import React from 'react'
 import { AiOutlineVideoCameraAdd } from 'react-icons/ai'
 import { HiOutlineStatusOnline, HiOutlineUpload } from 'react-icons/hi'
@@ -14,7 +14,7 @@ const NewVideoTrigger = () => {
   if (!selectedChannelId) return null
 
   return (
-    <Popover
+    <DropMenu
       trigger={
         <Button
           className="md:!block !hidden"
@@ -26,28 +26,31 @@ const NewVideoTrigger = () => {
     >
       <div className="p-1 mt-1.5 overflow-hidden border border-gray-100 rounded-xl shadow-xl dark:border-gray-800 bg-secondary">
         <div className="flex flex-col text-sm transition duration-150 ease-in-out rounded-md">
-          <Link
+          <Menu.Item
+            as={NextLink}
             href={UPLOAD}
             className="inline-flex items-center px-4 py-1.5 space-x-2 rounded-lg opacity-90 hover:opacity-100 hover:bg-gray-200 dark:hover:bg-gray-800"
           >
             <HiOutlineUpload />
             <span className="whitespace-nowrap">Upload Video</span>
-          </Link>
-          <button
-            type="button"
-            disabled
-            className="inline-flex opacity-40 items-center px-4 py-1.5 space-x-2 rounded-lg"
-          >
-            <Tooltip content="Coming soon">
-              <span className="inline-flex items-center space-x-2">
-                <HiOutlineStatusOnline className="text-red-500" />
-                <span className="whitespace-nowrap">Go Live Now</span>
-              </span>
-            </Tooltip>
-          </button>
+          </Menu.Item>
+          <Menu.Item>
+            <button
+              type="button"
+              disabled
+              className="inline-flex opacity-40 items-center px-4 py-1.5 space-x-2 rounded-lg"
+            >
+              <Tooltip content="Coming soon">
+                <span className="inline-flex items-center space-x-2">
+                  <HiOutlineStatusOnline className="text-red-500" />
+                  <span className="whitespace-nowrap">Go Live Now</span>
+                </span>
+              </Tooltip>
+            </button>
+          </Menu.Item>
         </div>
       </div>
-    </Popover>
+    </DropMenu>
   )
 }
