+++ src/components/Common/MoreTrigger.tsx
@@ -1,4 +1,5 @@
-import Popover from '@components/UIElements/Popover'
+import DropMenu, { NextLink } from '@components/UIElements/DropMenu'
+import { Menu } from '@headlessui/react'
 import {
   APP_NAME,
   LENSTUBE_GITHUB_HANDLE,
@@ -8,13 +9,12 @@ import {
 } from '@utils/constants'
 import { Mixpanel, TRACK } from '@utils/track'
 import { DISCORD, PRIVACY } from '@utils/url-path'
-import Link from 'next/link'
 import React from 'react'
 import { BsThreeDots } from 'react-icons/bs'
 
 const MoreTrigger = () => {
   return (
-    <Popover
+    <DropMenu
       trigger={
         <div className="flex flex-col space-y-2 mb-0.5">
           <button
@@ -35,99 +35,94 @@ const MoreTrigger = () => {
           <div className="text-[11px] cursor-default p-1 font-semibold uppercase opacity-50">
             More
           </div>
-          <div className="rounded-lg opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900">
-            <Link
-              href={LENSTUBE_STATUS_PAGE}
-              onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.STATUS)}
-              className="block px-2.5 py-1.5"
-              target="_blank"
-            >
-              Status
-            </Link>
-          </div>
-          <div className="rounded-lg opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900">
-            <Link
-              href={`https://github.com/${LENSTUBE_GITHUB_HANDLE}`}
-              onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.GITHUB)}
-              className="block px-2.5 py-1.5"
-              target="_blank"
-            >
-              Github
-            </Link>
-          </div>
-          <div className="rounded-lg opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900">
-            <Link
-              href={`https://twitter.com/${LENSTUBE_TWITTER_HANDLE}`}
-              onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.TWITTER)}
-              className="block px-2.5 py-1.5"
-              target="_blank"
-            >
-              Twitter
-            </Link>
-          </div>
-          <div className="rounded-lg opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900">
-            <Link
-              href={DISCORD}
-              onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.DISCORD)}
-              className="block px-2.5 py-1.5"
-              target="_blank"
-            >
-              Discord
-            </Link>
-          </div>
-          <div className="rounded-lg opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900">
-            <Link
-              href={`https://roadmap.lenstube.xyz`}
-              onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.ROADMAP)}
-              className="block px-2.5 py-1.5"
-              target="_blank"
-            >
-              Give Feedback
-            </Link>
-          </div>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg px-2.5 py-1.5 opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={LENSTUBE_STATUS_PAGE}
+            onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.STATUS)}
+            target="_blank"
+          >
+            Status
+          </Menu.Item>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg px-2.5 py-1.5 opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={`https://github.com/${LENSTUBE_GITHUB_HANDLE}`}
+            onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.GITHUB)}
+            target="_blank"
+          >
+            Github
+          </Menu.Item>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg px-2.5 py-1.5 opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={`https://twitter.com/${LENSTUBE_TWITTER_HANDLE}`}
+            onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.TWITTER)}
+            target="_blank"
+          >
+            Twitter
+          </Menu.Item>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg px-2.5 py-1.5 opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={DISCORD}
+            onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.DISCORD)}
+            target="_blank"
+          >
+            Discord
+          </Menu.Item>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg px-2.5 py-1.5 opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={`https://roadmap.lenstube.xyz`}
+            onClick={() => Mixpanel.track(TRACK.SYSTEM.MORE_MENU.ROADMAP)}
+            target="_blank"
+          >
+            Give Feedback
+          </Menu.Item>
           <div className="text-[11px] cursor-default p-1 font-semibold uppercase opacity-50">
             Legal
           </div>
-          <div className="rounded-lg hover:bg-gray-50 opacity-80 hover:opacity-100 dark:hover:bg-gray-900">
-            <Link href={PRIVACY} className="block px-2.5 py-1.5">
-              Privacy
-            </Link>
-          </div>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg px-2.5 py-1.5 opacity-80 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={PRIVACY}
+          >
+            Privacy
+          </Menu.Item>
           <div className="text-[11px] cursor-default p-1 font-semibold uppercase opacity-50">
             Powered by
           </div>
-          <div className="rounded-lg hover:bg-gray-50 opacity-80 hover:opacity-100 dark:hover:bg-gray-900">
-            <Link
-              href={`https://livepeer.studio/?utm_source=${APP_NAME}`}
-              rel="noreferrer noopener"
-              className="space-x-2 p-1.5 flex items-center"
-              target="_blank"
-            >
-              <span>
-                <img
-                  src={`${STATIC_ASSETS}/images/brand/lenstube.svg`}
-                  alt="lvpr"
-                  className="w-3 h-3"
-                  draggable={false}
-                />
-              </span>
-              <span>Livepeer</span>
-            </Link>
-          </div>
-          <div className="rounded-lg hover:bg-gray-50 opacity-80 hover:opacity-100 dark:hover:bg-gray-900">
-            <Link
-              href={`https://vercel.com/?utm_source=${APP_NAME}&utm_campaign=oss`}
-              rel="noreferrer noopener"
-              className="block space-x-2 p-1.5"
-              target="_blank"
-            >
-              <span>▲</span>
-              <span>Vercel</span>
-            </Link>
-          </div>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg opacity-80 space-x-2 p-1.5 flex items-center hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={`https://livepeer.studio/?utm_source=${APP_NAME}`}
+            rel="noreferrer noopener"
+            target="_blank"
+          >
+            <span>
+              <img
+                src={`${STATIC_ASSETS}/images/brand/lenstube.svg`}
+                alt="lvpr"
+                className="w-3 h-3"
+                draggable={false}
+              />
+            </span>
+            <span>Livepeer</span>
+          </Menu.Item>
+          <Menu.Item
+            as={NextLink}
+            className="rounded-lg opacity-80 space-x-2 p-1.5 hover:opacity-100 hover:bg-gray-50 dark:hover:bg-gray-900"
+            href={`https://vercel.com/?utm_source=${APP_NAME}&utm_campaign=oss`}
+            rel="noreferrer noopener"
+            target="_blank"
+          >
+            <span>▲</span>
+            <span>Vercel</span>
+          </Menu.Item>
         </div>
       </div>
-    </Popover>
+    </DropMenu>
   )
 }
