+++ src/components/Notifications/Subscribed.tsx
@@ -1,5 +1,7 @@
 import IsVerified from '@components/Common/IsVerified'
 import AddressExplorerLink from '@components/Common/Links/AddressExplorerLink'
+import { NextLink } from '@components/UIElements/DropMenu'
+import { Menu } from '@headlessui/react'
 import useAppStore from '@lib/store'
 import getProfilePicture from '@utils/functions/getProfilePicture'
 import { getRandomProfilePicture } from '@utils/functions/getRandomProfilePicture'
@@ -7,7 +9,6 @@ import imageCdn from '@utils/functions/imageCdn'
 import { shortenAddress } from '@utils/functions/shortenAddress'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import Link from 'next/link'
 import React, { FC } from 'react'
 import { NewFollowerNotification } from 'src/types'
 
@@ -24,7 +25,8 @@ const SubscribedNotification: FC<Props> = ({ notification }) => {
     <>
       <div className="flex items-center space-x-3">
         {notification?.wallet?.defaultProfile ? (
-          <Link
+          <Menu.Item
+            as={NextLink}
             href={`/${notification?.wallet?.defaultProfile?.handle}`}
             className="inline-flex items-center space-x-1.5 font-base"
           >
@@ -44,7 +46,7 @@ const SubscribedNotification: FC<Props> = ({ notification }) => {
                 size="xs"
               />
             </div>
-          </Link>
+          </Menu.Item>
         ) : (
           <AddressExplorerLink address={notification?.wallet?.address}>
             <span className="inline-flex items-center space-x-1.5 font-base">
